added shortcode template for expandable sections
Added a new shortcode template that will allow you to have a collapsible
div on the page. This div can be clicked to open and show the content
on the page, and then clicked again to have the content be hidden.

Example on how to use the shortcode block:

```markdown
{{< details "Click to expand this section for new tests" >}}
Something cool here....
{{< /details >}}
```

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>